refactor(auth): move permission catalogue to shared, add arch guard (#938)
shared/auth/permissions.py (and shared/web/deps, scope_catalog, endpoint_scopes)
imported the implication logic from admin.core.permissions — a shared→admin
layering inversion (admin already imports shared: a latent cycle) that no arch
guard prevented.

The permission catalogue has no admin-tier dependency (only shared/scopes.py), so
move Permission / ALL_PERMISSIONS / IMPLICATION_MAP / compute_effective /
compute_implies_transitive to a new tier-neutral shared.auth.permission_catalog.
admin.core.permissions becomes a thin re-export shim (explicit __all__), so every
existing admin-tier call site keeps working unchanged; the four shared importers
now point at the shared module directly.

Add test_shared_does_not_import_admin_permissions, scoped to the permissions
module: shared jobs/events/audit still reference admin-tier ORM schemas (a
separate pre-existing coupling, out of scope for #938), but the permissions
inversion this issue names is now gone and guarded.
